+++ lib/flags.ml
@@ -53,8 +53,6 @@ let in_toplevel = ref false
 
 let profile = false
 
-let raw_print = ref false
-
 let we_are_parsing = ref false
 
 (* Compatibility mode *)