+++ lib/flags.mli
@@ -49,9 +49,6 @@ val profile : bool
 (* development flag to detect race conditions, it should go away. *)
 val we_are_parsing : bool ref
 
-(* Set Printing All flag. For some reason it is a global flag *)
-val raw_print : bool ref
-
 type compat_version = V8_7 | V8_8 | V8_9 | Current
 val compat_version : compat_version ref
 val version_compare : compat_version -> compat_version -> int