+++ stm/vernac_classifier.ml
@@ -159,6 +159,8 @@ let classify_vernac e =
     | VernacUnsetOption _ | VernacSetOption _
     | VernacAddOption _ | VernacRemoveOption _
     | VernacMemOption _ | VernacPrintOption _
+    | VernacSetPrintingAll | VernacSetPrintingSugared
+    | VernacSetPrintingDefaults | VernacUnsetPrintingAll
     | VernacGlobalCheck _
     | VernacDeclareReduction _
     | VernacExistingClass _ | VernacExistingInstance _