+++ vernac/explainErr.ml
@@ -50,9 +50,9 @@ let wrap_vernac_error (exn, info) strm = (EvaluatedError (strm, None), info)
 
 let process_vernac_interp_error exn = match fst exn with
   | Univ.UniverseInconsistency i ->
-    let msg =
-      if !Constrextern.print_universes then
-	str "." ++ spc() ++ 
+     let msg =
+       if Printoptions.((get()).universes) then
+        str "." ++ spc() ++
           Univ.explain_universe_inconsistency UnivNames.pr_with_global_universes i
       else
 	mt() in